Extract FusionMode and FirstEstimate to types.py to avoid circular dependency
- New types.py holds both enums, imported by backends, deconvolver, estimates
- backends.py uses FusionMode.SUMMATIVE enum comparison (not .value string)
- deconvolver.py and estimates.py import from types.py, no circular import
- 396 tests pass
